separate inventory usecase interactor tests into separate files
