--- src/test/java/use_case/inventory/InventoryInteractorTest.java
@@ -1,192 +1,0 @@
-package use_case.inventory;
-
-import java.util.Map;
-
-import org.junit.jupiter.api.Test;
-
-import data_access.InMemoryInventoryDataAccessObject;
-import entity.AbstractFood;
-import entity.FoodInventoryFactory;
-import entity.FoodItemFactory;
-import entity.Inventory;
-import entity.InventoryFactory;
-import use_case.food_management.add_to_inventory.AddToInventoryInputBoundary;
-import use_case.food_management.add_to_inventory.AddToInventoryInputData;
-import use_case.food_management.add_to_inventory.AddToInventoryInteractor;
-import use_case.food_management.add_to_inventory.AddToInventoryOutputBoundary;
-import use_case.food_management.add_to_inventory.AddToInventoryOutputData;
-import use_case.food_management.create_inventory.CreateInventoryInputBoundary;
-import use_case.food_management.create_inventory.CreateInventoryInputData;
-import use_case.food_management.create_inventory.CreateInventoryInteractor;
-import use_case.food_management.create_inventory.CreateInventoryOutputBoundary;
-import use_case.food_management.create_inventory.CreateInventoryOutputData;
-import use_case.food_management.use_item_in_inventory.UseItemInputBoundary;
-import use_case.food_management.use_item_in_inventory.UseItemInputData;
-import use_case.food_management.use_item_in_inventory.UseItemInteractor;
-import use_case.food_management.use_item_in_inventory.UseItemOutputBoundary;
-import use_case.food_management.use_item_in_inventory.UseItemOutputData;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-class InventoryInteractorTest {
-
-    @Test
-    void addToEmptyInventoryTest() {
-        final AddToInventoryInputData inputData = new AddToInventoryInputData("chiually", "milk");
-        final InMemoryInventoryDataAccessObject inventoryRepository = new InMemoryInventoryDataAccessObject();
-
-        final InventoryFactory inventoryFactory = new FoodInventoryFactory();
-        final FoodItemFactory foodItemFactory = new FoodItemFactory();
-        // create inventory for testing
-        final Inventory inventory = inventoryFactory.create("chiually");
-        inventoryRepository.save(inventory);
-
-        final AddToInventoryOutputBoundary successPresenter = new AddToInventoryOutputBoundary() {
-            @Override
-            public void prepareSuccessView(AddToInventoryOutputData inventory) {
-                // check that output data contains username and food id
-                assertTrue(inventory.isSuccess());
-            }
-        };
-
-        final AddToInventoryInputBoundary interactor = new AddToInventoryInteractor(inventoryRepository,
-                successPresenter, foodItemFactory);
-        interactor.execute(inputData);
-
-    }
-
-    @Test
-    void addToInventoryWithItemTest() {
-        final InMemoryInventoryDataAccessObject inventoryRepository = new InMemoryInventoryDataAccessObject();
-        final AddToInventoryInputData inputData = new AddToInventoryInputData("chiually", "milk");
-        final InventoryFactory inventoryFactory = new FoodInventoryFactory();
-        final FoodItemFactory foodItemFactory = new FoodItemFactory();
-
-        // create inventory for user
-        final Inventory inventory = inventoryFactory.create("chiually");
-        // add same food item
-        final Map<String, AbstractFood> inventoryItems = inventory.getItems();
-        inventoryItems.put("milk", foodItemFactory.create("milk"));
-        inventory.setItems(inventoryItems);
-        inventoryRepository.save(inventory);
-
-        final AddToInventoryOutputBoundary successPresenter = new AddToInventoryOutputBoundary() {
-
-            @Override
-            public void prepareSuccessView(AddToInventoryOutputData inventory) {
-                assertTrue(inventory.isSuccess());
-            }
-        };
-        final AddToInventoryInputBoundary interactor = new AddToInventoryInteractor(inventoryRepository,
-                successPresenter, foodItemFactory);
-        interactor.execute(inputData);
-    }
-
-    @Test
-    void useLastItemTest() {
-        final UseItemInputData inputData = new UseItemInputData("chiually", "milk");
-        final InMemoryInventoryDataAccessObject inventoryRepository = new InMemoryInventoryDataAccessObject();
-
-        final InventoryFactory inventoryFactory = new FoodInventoryFactory();
-        final FoodItemFactory foodItemFactory = new FoodItemFactory();
-
-        final Inventory inventory = inventoryFactory.create("chiually");
-        final AbstractFood foodItem = foodItemFactory.create("milk");
-        foodItem.setQuantity(1);
-        final Map<String, AbstractFood> inventoryItems = inventory.getItems();
-        inventoryItems.put("milk", foodItem);
-        inventory.setItems(inventoryItems);
-        inventoryRepository.save(inventory);
-
-        final UseItemOutputBoundary successPresenter = new UseItemOutputBoundary() {
-
-            @Override
-            public void prepareSuccessView(UseItemOutputData useInventoryOutputData) {
-                assertTrue(useInventoryOutputData.isSuccess());
-            }
-        };
-        final UseItemInputBoundary interactor = new UseItemInteractor(inventoryRepository, successPresenter);
-        interactor.execute(inputData);
-    }
-
-    @Test
-    void useOneOfMultipleItemTest() {
-        final UseItemInputData inputData = new UseItemInputData("chiually", "milk");
-        final InMemoryInventoryDataAccessObject inventoryRepository = new InMemoryInventoryDataAccessObject();
-
-        final InventoryFactory inventoryFactory = new FoodInventoryFactory();
-        final FoodItemFactory foodItemFactory = new FoodItemFactory();
-
-        final Inventory inventory = inventoryFactory.create("chiually");
-        final AbstractFood foodItem = foodItemFactory.create("milk");
-        foodItem.setQuantity(2);
-        final Map<String, AbstractFood> inventoryItems = inventory.getItems();
-        inventoryItems.put("milk", foodItem);
-        inventory.setItems(inventoryItems);
-        inventoryRepository.save(inventory);
-
-        final UseItemOutputBoundary successPresenter = new UseItemOutputBoundary() {
-
-            @Override
-            public void prepareSuccessView(UseItemOutputData useInventoryOutputData) {
-                assertTrue(useInventoryOutputData.isSuccess());
-            }
-        };
-        final UseItemInputBoundary interactor = new UseItemInteractor(inventoryRepository, successPresenter);
-        interactor.execute(inputData);
-    }
-
-    @Test
-    void createNonExistentInventoryTest() {
-        final CreateInventoryInputData inputData = new CreateInventoryInputData("chiually");
-        final InMemoryInventoryDataAccessObject inventoryRepository = new InMemoryInventoryDataAccessObject();
-
-        final InventoryFactory inventoryFactory = new FoodInventoryFactory();
-
-        final CreateInventoryOutputBoundary successPresenter = new CreateInventoryOutputBoundary() {
-
-            @Override
-            public void prepareSuccessView(CreateInventoryOutputData inventory) {
-                assertTrue(inventory.isSuccess());
-            }
-        };
-
-        final CreateInventoryInputBoundary interactor = new CreateInventoryInteractor(inventoryRepository,
-                successPresenter, inventoryFactory);
-        interactor.execute(inputData);
-    }
-
-    @Test
-    void createExistingInventoryTest() {
-        final CreateInventoryInputData inputData = new CreateInventoryInputData("chiually");
-        final InMemoryInventoryDataAccessObject inventoryRepository = new InMemoryInventoryDataAccessObject();
-
-        final InventoryFactory inventoryFactory = new FoodInventoryFactory();
-        final FoodItemFactory foodItemFactory = new FoodItemFactory();
-
-        final Inventory inventory = inventoryFactory.create("chiually");
-        final AbstractFood foodItem = foodItemFactory.create("milk");
-        foodItem.setQuantity(2);
-        final Map<String, AbstractFood> inventoryItems = inventory.getItems();
-        inventoryItems.put("milk", foodItem);
-        inventory.setItems(inventoryItems);
-        inventoryRepository.save(inventory);
-
-        final CreateInventoryOutputBoundary successPresenter = new CreateInventoryOutputBoundary() {
-
-            @Override
-            public void prepareSuccessView(CreateInventoryOutputData inventory) {
-
-                assertTrue(inventory.isSuccess());
-                assertTrue(inventory.getInventoryItems().containsKey("milk"));
-                assertEquals(2, inventory.getInventoryItems().get("milk").getQuantity());
-            }
-        };
-
-        final CreateInventoryInputBoundary interactor = new CreateInventoryInteractor(inventoryRepository,
-                successPresenter, inventoryFactory);
-        interactor.execute(inputData);
-    }
-
-}
